+++ Krowi_AchievementFilter/Data/Data.lua
@@ -6,6 +6,19 @@ local data = addon.Data;
 addon.Categories = {};
 addon.Achievements = {};
 
+function data.Load()
+    addon.Classic.Load();
+    addon.TheBurningCrusade.Load();
+    addon.WrathOfTheLichKing.Load();
+    addon.Cataclysm.Load();
+    addon.MistsOfPandaria.Load();
+    addon.WarlordsOfDreanor.Load();
+    addon.Legion.Load();
+    addon.BattleForAzeroth.Load();
+    addon.Shadowlands.Load();
+    addon.Categories, addon.Achievements = addon.Data:GetLists(); -- This freezes the game if we load it outside of the loading screen
+end
+
 local function ConvertToAchievementFrameCategory(datum, categories, achievements)
     if datum.Level == 0 then
         datum.NotHidden = true;