+++ Krowi_AchievementFilter/Tutorials.lua
@@ -13,6 +13,7 @@ local function GetTitle(title)
     return AF_COLOR_YELLOW .. title .. AF_COLOR_END .. "\n\n";
 end
 
+local tabButton1, categoriesFrame, achievementsFrame, searchBoxFrame, searchPreviewFrame, fullSearchResultsFrame;
 function tutorials.Load()
     tutorials.RegisterTutorial(tutorials.FeaturesTutorial, {
         savedvariable = addon.Options.db,
@@ -80,51 +81,62 @@ function tutorials.Load()
             shineRight = 13,
         },
         onShow = function(self, i)
+            local gui = addon.GUI;
+
             if i == 1 then
-                addon.OpenAchievementFrameAtTabButton1();
-                self[i].shine = addon.GUI.TabButton1;
+                gui.ToggleAchievementFrameAtTab1();
+                self[i].shine = tabButton1;
             elseif i == 2 then
-                addon.OpenAchievementFrameAtTabButton1();
-                addon.GUI.CategoriesFrame:SelectCategory(addon.GetAchievement(14281):GetCategory());
-                self[i].shine = addon.GUI.CategoriesFrame;
+                gui.ToggleAchievementFrameAtTab1();
+                categoriesFrame:SelectCategory(addon.GetAchievement(14281):GetCategory());
+                self[i].shine = categoriesFrame;
             elseif i == 3 then
-                addon.OpenAchievementFrameAtTabButton1();
-                local achievementsButtons = addon.GUI.AchievementsFrame.Container.buttons;
-                addon.GUI.AchievementsFrame:SelectAchievementFromID(1283, "RightButton", true, achievementsButtons[1], 88, 34);
+                gui.ToggleAchievementFrameAtTab1();
+                local achievementsButtons = achievementsFrame.Container.buttons;
+                achievementsFrame:SelectAchievementFromID(1283, "RightButton", true, achievementsButtons[1], 88, 34);
                 self[i].shine = DropDownList1;
             elseif i == 4 then
-                addon.OpenAchievementFrameAtTabButton1();
-                addon.GUI.SearchBoxFrame:SetText("cla");
-                addon.GUI.SearchBoxFrame:OnTextChanged(); -- Trigger this one manually as the previous line does not trigger it in order to search for achievements
+                gui.ToggleAchievementFrameAtTab1();
+                searchBoxFrame:SetText("cla");
+                searchBoxFrame:OnTextChanged(); -- Trigger this one manually as the previous line does not trigger it in order to search for achievements
                 local bottom;
-                if addon.GUI.SearchPreviewFrame.ShowFullSearchResultsButton:IsShown() then
-                    bottom = addon.GUI.SearchPreviewFrame.ShowFullSearchResultsButton:GetBottom();
+                if searchPreviewFrame.ShowFullSearchResultsButton:IsShown() then
+                    bottom = searchPreviewFrame.ShowFullSearchResultsButton:GetBottom();
                 else
-                    local buttons = addon.GUI.SearchPreviewFrame.Buttons;
+                    local buttons = searchPreviewFrame.Buttons;
                     for _, button in next, buttons do
                         if button:IsShown() then
                             bottom = button:GetBottom();
                         end
                     end
                 end
-                self[i].shineBottom = bottom - addon.GUI.SearchPreviewFrame:GetBottom() - 10;
-                self[i].shine = addon.GUI.SearchPreviewFrame;
+                self[i].shineBottom = bottom - searchPreviewFrame:GetBottom() - 10;
+                self[i].shine = searchPreviewFrame;
             elseif i == 5 then
-                addon.OpenAchievementFrameAtTabButton1();
-                addon.GUI.SearchBoxFrame:SetText("cla");
-                addon.GUI.SearchBoxFrame:OnTextChanged(); -- Trigger this one manually as the previous line does not trigger it in order to search for achievements
-                addon.GUI.SearchPreviewFrame.ShowFullSearchResultsButton:Click();
-                addon.GUI.SearchBoxFrame:SetText("");
-                addon.GUI.SearchBoxFrame:OnTextChanged();
-                self[i].shine = addon.GUI.FullSearchResultsFrame;
+                gui.ToggleAchievementFrameAtTab1();
+                searchBoxFrame:SetText("cla");
+                searchBoxFrame:OnTextChanged(); -- Trigger this one manually as the previous line does not trigger it in order to search for achievements
+                searchPreviewFrame.ShowFullSearchResultsButton:Click();
+                searchBoxFrame:SetText("");
+                searchBoxFrame:OnTextChanged();
+                self[i].shine = fullSearchResultsFrame;
             end
         end
     });
     tutorials.CloseButtonHook(tutorials.FeaturesTutorial, function()
         diagnostics.Trace("tutorials.CloseButtonHook");
 
-        addon.ResetView(addon.GUI.CategoriesFrame, addon.GUI.SearchBoxFrame, addon.GUI.FullSearchResultsFrame);
+        addon.GUI.ResetView();
     end);
+end
+
+function tutorials.SetFrames(tabBtn1, catFrame, achFrame, srchBxFrame, srchPrvwFrame, fullSrchRsltsFrame)
+    tabButton1 = tabBtn1;
+    categoriesFrame = catFrame;
+    achievementsFrame = achFrame;
+    searchBoxFrame = srchBxFrame;
+    searchPreviewFrame = srchPrvwFrame;
+    fullSearchResultsFrame = fullSrchRsltsFrame;
 end
 
 function tutorials.HookTrigger(hook)